build-up-server/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.2.1
fastapi==0.110.0
h11==0.16.0
idna==3.10
psutil==7.0.0
pydantic==2.11.5
pydantic_core==2.33.2
PyJWT==2.8.0
sniffio==1.3.1
starlette==0.36.3
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn==0.27.1
